requirements.txt
django
mysqlclient
python-dotenv
django-bootstrap4
django-bootstrap-datepicker-plus